Create factory for order and its nested fields
